# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "0.1"
    },
    "paths": {
...
        "/api/region": {
            "get": {
                "tags": [
                    "Region"
                ],
                "summary": "Get a list of region with dynamic filtering, sorting, pagination, and advanced search",
                "operationId": "0c51f58df055e5721d4627d15b7ceb04",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Dynamic filtering with multiple fields (e.g., `status:active|age:gt:30`)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Full-text search (e.g., `keywords`)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort the results by a specific field (e.g., `price` or `name`)",
                        "schema": {
                            "type": "string",
                            "example": "created_at"
                        }
                    },
                    {
                        "name": "sort_dir",
                        "in": "query",
                        "description": "Direction of sorting (asc or desc)",
                        "schema": {
                            "type": "string",
                            "default": "asc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination (default: 1)",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page (default: 10)",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/RegionResource"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "next_page_url": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "prev_page_url": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Region"
                ],
                "summary": "Store a new region",
                "operationId": "ecd6f6bccf8409fa68563eeb8dad122d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Region created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RegionResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/region/{id}": {
            "get": {
                "tags": [
                    "Region"
                ],
                "summary": "Get a specific region",
                "operationId": "936672c6b1827058659d1796787c13c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Region ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RegionResource"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Region"
                ],
                "summary": "Update a specific region",
                "operationId": "7141803d2fcfcf13c52928729e78ad03",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Region ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Region updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RegionResource"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Region"
                ],
                "summary": "Delete a region",
                "operationId": "3982e32c404e6a6446ce94ac638b3881",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Region ID",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Region deleted successfully"
                    }
                }
            }
        },
        "/api/region/{id}/restore": {
            "put": {
                "tags": [
                    "Region"
                ],
                "summary": "Restore a deleted region",
                "operationId": "860a617f673317a574bc689a3d517d03",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Region ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Region restored successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The record has been restored successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/region/{id}/force-destroy": {
            "delete": {
                "tags": [
                    "Region"
                ],
                "summary": "Permanently delete a region",
                "operationId": "9d25030266bc6bf35caa324b23f05e71",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Region ID",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Region permanently deleted"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RegionRequest": {
                "required": [
                    ""
                ],
                "properties": {
                    "name_en": {
                        "type": "string",
                        "example": "Name_en Example"
                    },
                    "name_mm": {
                        "type": "string",
                        "example": "Name_mm Example"
                    },
                    "code": {
                        "type": "string",
                        "example": "Code Example"
                    }
                },
                "type": "object"
            },
...
            "RegionResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "62"
                    },
                    "name_en": {
                        "type": "string",
                        "example": "Name_en Example"
                    },
                    "name_mm": {
                        "type": "string",
                        "example": "Name_mm Example"
                    },
                    "code": {
                        "type": "string",
                        "example": "Code Example"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-31 09:12:31"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-31 09:12:31"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token Based Authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}